stock-analysis/packages/ai/analysis/requirements.txt
beautifulsoup4==4.13.4
openai==1.93.0
Requests==2.32.4

python-dotenv==1.1.1